fixes for 64-bit support

+++ JSVM/H264Extension/include/H264AVCCommonIf.h
@@ -55,7 +55,7 @@ typedef MemAccessor< UChar > BinDataAccessor;
 
 typedef UChar   Pel;
 typedef Short   XPel;
-typedef Long    XXPel;
+typedef Int     XXPel;
 
 class TCoeff
 {
